refactor :json 저장 함수 목적에 맞추어 file_operations으로 수정
