+++ airflow/scripts/utils/api_operations.py
@@ -5,7 +5,6 @@ import os
 from io import StringIO
 import requests
 import time
-import json
 import csv
 
 
@@ -100,7 +99,7 @@ def get_headers(site: str, key_num: int) -> Dict[str, str]:
 
 def save_csv_file(file_path: str, isbn_list: List[str]) -> None:
     """
-    ISBN 목록을 주어진 경로의 CSV 파일로 저장합니다.
+    ISBN 목록을 주어진 경로의 CSV 파일로 저장합니다. (init이 아닌 데이터 로직에서 사용하는 save_csv_file은 목적에 따라 file_operations로 이동)
 
     Args:
         file_path (str): ISBN 목록을 저장할 파일 경로
@@ -116,23 +115,6 @@ def save_csv_file(file_path: str, isbn_list: List[str]) -> None:
         writer.writerow(["ISBN"])
         for isbn in isbn_list:
             writer.writerow([isbn])
-
-
-def save_json_file(file_path: str, items: Dict[str, Dict]) -> None:
-    """
-    주어진 항목을 JSON 파일로 저장합니다.
-
-    Args:
-        file_path (str): 항목을 저장할 파일 경로
-        items (list or dict): JSON 형식으로 저장할 항목
-    """
-    directory = os.path.dirname(file_path)
-
-    if not os.path.exists(directory):
-        os.makedirs(directory)
-
-    with open(file_path, "w", encoding="utf-8") as f:
-        json.dump(items, f, ensure_ascii=False, indent=4)
 
 
 def fetch_api_data(isbn_list: List[str], site: str) -> Dict[str, Dict]:
